TokenServiceに、 １、ユーザー登録時にトークンを発行するメソッドを追加 ２、ログイン時、idとパスワードが一致しているを確認するための分岐を作成

+++ api/src/main/kotlin/com/kyotob/api/service/TokenService.kt
@@ -1,25 +1,35 @@
 package com.kyotob.api.service
 
-import org.springframework.stereotype.Service
 import com.kyotob.api.controller.UnauthorizedException
-import com.kyotob.api.model.Token
 import com.kyotob.api.mapper.TokenDao
+import com.kyotob.api.model.Token
+import org.springframework.stereotype.Service
 import java.util.*
 
 
 @Service
-class TokenService(private val tokenDao: TokenDao) {
-
-    fun getAccessToken(userId: Int, password: String): String {
-        //Todo: idとパスワードが正しいことを確認
-
-        //
-        tokenDao.delete(userId)
-        //Todo: いい感じのトークンを返す
-        //とりあえずuserIdを文字列にしたやつをハッシュ化しておく
+class TokenService(private val tokenDao: TokenDao, private val userService: UserService) {
+    fun createAccessToken(userId: Int): String{
         val token: String = UUID.randomUUID().toString()
         tokenDao.create(userId, token)
         return token
+    }
+
+    fun getAccessToken(userId: Int, password: String): String {
+        //idとパスワードが正しいことを確認
+        if(!userService.isConsistent(userId, password)){
+            //idとパスワードが一致しなければエラーを投げる
+            throw UnauthorizedException("id and password not match")
+        }
+
+        else {
+            //一致すれば新しいtokenを発行
+            tokenDao.delete(userId)
+            //とりあえずuserIdを文字列にしたやつをハッシュ化しておく
+            val token: String = UUID.randomUUID().toString()
+            tokenDao.create(userId, token)
+            return token
+        }
     }
 
     fun verifyAccessToken(accessToken: String): Int {
